streaming/requirements.txt
kafka-python==2.0.2
faker==19.12.0
numpy
requests